# Restructure instrument creation code paths