--- a/sdk/metric/pipeline_registry_test.go
+++ b/sdk/metric/pipeline_registry_test.go
@@ -21,7 +21,7 @@ import (
 	"github.com/stretchr/testify/require"
 
 	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/metric/unit"
+	"go.opentelemetry.io/otel/sdk/instrumentation"
 	"go.opentelemetry.io/otel/sdk/metric/aggregation"
 	"go.opentelemetry.io/otel/sdk/metric/internal"
 	"go.opentelemetry.io/otel/sdk/metric/view"
@@ -57,7 +57,7 @@ func testCreateAggregators[N int64 | float64](t *testing.T) {
 		view.WithSetAggregation(invalidAggregation{}),
 	)
 
-	instruments := []view.Instrument{
+	instruments := []instProviderKey{
 		{Name: "foo", Kind: view.InstrumentKind(0)}, //Unknown kind
 		{Name: "foo", Kind: view.SyncCounter},
 		{Name: "foo", Kind: view.SyncUpDownCounter},
@@ -71,7 +71,7 @@ func testCreateAggregators[N int64 | float64](t *testing.T) {
 		name     string
 		reader   Reader
 		views    []view.View
-		inst     view.Instrument
+		inst     instProviderKey
 		wantKind internal.Aggregator[N] //Aggregators should match len and types
 		wantLen  int
 		wantErr  error
@@ -209,10 +209,11 @@ func testCreateAggregators[N int64 | float64](t *testing.T) {
 			wantErr: errCreatingAggregators,
 		},
 	}
+	s := instrumentation.Scope{Name: "testCreateAggregators"}
 	for _, tt := range testcases {
 		t.Run(tt.name, func(t *testing.T) {
-			i := newInserter[N](newPipeline(nil, tt.reader, tt.views))
-			got, err := i.Instrument(tt.inst, unit.Dimensionless)
+			i := newInserter[N](s, newPipeline(nil, tt.reader, tt.views))
+			got, err := i.Instrument(tt.inst)
 			assert.ErrorIs(t, err, tt.wantErr)
 			require.Len(t, got, tt.wantLen)
 			for _, agg := range got {
@@ -223,12 +224,13 @@ func testCreateAggregators[N int64 | float64](t *testing.T) {
 }
 
 func testInvalidInstrumentShouldPanic[N int64 | float64]() {
-	i := newInserter[N](newPipeline(nil, NewManualReader(), []view.View{{}}))
-	inst := view.Instrument{
+	s := instrumentation.Scope{Name: "testInvalidInstrumentShouldPanic"}
+	i := newInserter[N](s, newPipeline(nil, NewManualReader(), []view.View{{}}))
+	inst := instProviderKey{
 		Name: "foo",
 		Kind: view.InstrumentKind(255),
 	}
-	_, _ = i.Instrument(inst, unit.Dimensionless)
+	_, _ = i.Instrument(inst)
 }
 
 func TestInvalidInstrumentShouldPanic(t *testing.T) {
@@ -332,20 +334,22 @@ func TestPipelineRegistryCreateAggregators(t *testing.T) {
 }
 
 func testPipelineRegistryResolveIntAggregators(t *testing.T, p pipelines, wantCount int) {
-	inst := view.Instrument{Name: "foo", Kind: view.SyncCounter}
+	inst := instProviderKey{Name: "foo", Kind: view.SyncCounter}
 
-	r := newResolver[int64](p)
-	aggs, err := r.Aggregators(inst, unit.Dimensionless)
+	s := instrumentation.Scope{Name: "testPipelineRegistryResolveIntAggregators"}
+	r := newResolver[int64](s, p)
+	aggs, err := r.Aggregators(inst)
 	assert.NoError(t, err)
 
 	require.Len(t, aggs, wantCount)
 }
 
 func testPipelineRegistryResolveFloatAggregators(t *testing.T, p pipelines, wantCount int) {
-	inst := view.Instrument{Name: "foo", Kind: view.SyncCounter}
+	inst := instProviderKey{Name: "foo", Kind: view.SyncCounter}
 
-	r := newResolver[float64](p)
-	aggs, err := r.Aggregators(inst, unit.Dimensionless)
+	s := instrumentation.Scope{Name: "testPipelineRegistryResolveFloatAggregators"}
+	r := newResolver[float64](s, p)
+	aggs, err := r.Aggregators(inst)
 	assert.NoError(t, err)
 
 	require.Len(t, aggs, wantCount)
@@ -373,17 +377,18 @@ func TestPipelineRegistryCreateAggregatorsIncompatibleInstrument(t *testing.T) {
 		},
 	}
 	p := newPipelines(resource.Empty(), views)
-	inst := view.Instrument{Name: "foo", Kind: view.AsyncGauge}
+	inst := instProviderKey{Name: "foo", Kind: view.AsyncGauge}
 
-	ri := newResolver[int64](p)
-	intAggs, err := ri.Aggregators(inst, unit.Dimensionless)
+	s := instrumentation.Scope{Name: "TestPipelineRegistryCreateAggregatorsIncompatibleInstrument"}
+	ri := newResolver[int64](s, p)
+	intAggs, err := ri.Aggregators(inst)
 	assert.Error(t, err)
 	assert.Len(t, intAggs, 0)
 
 	p = newPipelines(resource.Empty(), views)
 
-	rf := newResolver[float64](p)
-	floatAggs, err := rf.Aggregators(inst, unit.Dimensionless)
+	rf := newResolver[float64](s, p)
+	floatAggs, err := rf.Aggregators(inst)
 	assert.Error(t, err)
 	assert.Len(t, floatAggs, 0)
 }
@@ -400,33 +405,34 @@ func TestPipelineRegistryCreateAggregatorsDuplicateErrors(t *testing.T) {
 		},
 	}
 
-	fooInst := view.Instrument{Name: "foo", Kind: view.SyncCounter}
-	barInst := view.Instrument{Name: "bar", Kind: view.SyncCounter}
+	fooInst := instProviderKey{Name: "foo", Kind: view.SyncCounter}
+	barInst := instProviderKey{Name: "bar", Kind: view.SyncCounter}
 
 	p := newPipelines(resource.Empty(), views)
 
-	ri := newResolver[int64](p)
-	intAggs, err := ri.Aggregators(fooInst, unit.Dimensionless)
+	s := instrumentation.Scope{Name: "TestPipelineRegistryCreateAggregatorsDuplicateErrors"}
+	ri := newResolver[int64](s, p)
+	intAggs, err := ri.Aggregators(fooInst)
 	assert.NoError(t, err)
 	assert.Len(t, intAggs, 1)
 
 	// The Rename view should error, because it creates a foo instrument.
-	intAggs, err = ri.Aggregators(barInst, unit.Dimensionless)
+	intAggs, err = ri.Aggregators(barInst)
 	assert.Error(t, err)
 	assert.Len(t, intAggs, 2)
 
 	// Creating a float foo instrument should error because there is an int foo instrument.
-	rf := newResolver[float64](p)
-	floatAggs, err := rf.Aggregators(fooInst, unit.Dimensionless)
+	rf := newResolver[float64](s, p)
+	floatAggs, err := rf.Aggregators(fooInst)
 	assert.Error(t, err)
 	assert.Len(t, floatAggs, 1)
 
-	fooInst = view.Instrument{Name: "foo-float", Kind: view.SyncCounter}
+	fooInst = instProviderKey{Name: "foo-float", Kind: view.SyncCounter}
 
-	_, err = rf.Aggregators(fooInst, unit.Dimensionless)
+	_, err = rf.Aggregators(fooInst)
 	assert.NoError(t, err)
 
-	floatAggs, err = rf.Aggregators(barInst, unit.Dimensionless)
+	floatAggs, err = rf.Aggregators(barInst)
 	assert.Error(t, err)
 	assert.Len(t, floatAggs, 2)
 }