### `instProvider` type

Currently, instrument providers are defined for the async and sync instruments both of `int64` and `float64` types. That produces four instrument providers in total. Each instrument provider has three methods to create three different instruments. Each of these methods does the following:

1. Create an instrument configuration
2. Create a `view.Instrument` with the appropriate `view.InstrumentKind` based on the instrument provider being used.
3. Resolve aggregators
4. Validate the return of that resolution
5. Return the instrument and error

For example,

https://github.com/open-telemetry/opentelemetry-go/blob/3675379e73448c28742f6e8c02c8be84c13c2bf9/sdk/metric/instrument_provider.go#L36-L53

The only functionally unique part of these methods is in step 2 (step 5 technically involves a cast to an API type, but it ends up not being relevant).

These common functional elements can be abstracted into a common method. However, this can be taken further. Given each instrument provider would do this, we can define a common underlying type for all instrument providers that defines this method. E.g.

```go
type instProvider[N int64 | float64] struct {
	resolve resolver[N]
}

func (p instProvider[N]) lookup(kind view.InstrumentKind, name string, opts []instrument.Option) (*instrumentImpl[N], error) {
	cfg := instrument.NewConfig(opts...)
	key := instProviderKey{
		Name:        name,
		Description: cfg.Description(),
		Unit:        cfg.Unit(),
		Kind:        kind,
	}

	aggs, err := p.resolve.Aggregators(key)
	if len(aggs) == 0 && err != nil {
		err = fmt.Errorf("instrument does not match any view: %w", err)
	}
	return &instrumentImpl[N]{aggregators: aggs}, err
}
```

Doing so, allows each instrument provider to be reduced to something resembling this:

```go
type asyncInt64Provider struct {
	instProvider[int64]
}

func (p asyncInt64Provider) Counter(name string, opts ...instrument.Option) (asyncint64.Counter, error) {
	return p.lookup(view.AsyncCounter, name, opts)
}
func (p asyncInt64Provider) UpDownCounter(name string, opts ...instrument.Option) (asyncint64.UpDownCounter, error) {
	return p.lookup(view.AsyncUpDownCounter, name, opts)
}
func (p asyncInt64Provider) Gauge(name string, opts ...instrument.Option) (asyncint64.Gauge, error) {
	return p.lookup(view.AsyncGauge, name, opts)
}
```

This both reduces a lot of redundant code and centralizes the functional section development needs to deal with going forward.

### Store instrument providers in a meter

Currently, when a `meter` is asked for a new instrument provider it creates a new one every time:

https://github.com/open-telemetry/opentelemetry-go/blob/3675379e73448c28742f6e8c02c8be84c13c2bf9/sdk/metric/meter.go#L92-L117

This includes allocating a pointer for a new resolver. An allocation is unavoidable given the resolver contains a slice and even if passed as a value, the new reference to a slice would incur an allocation.

As an alternative to this approach, allocate once-per-provider and do so at the start.

#### Performance considerations

Using the included benchmark for the change we can see this indeed reduces the overall allocations for new instrument creation:

```
name                      old time/op    new time/op    delta
InstrumentCreation-8        23.8µs ± 3%    21.4µs ± 4%  -10.13%  (p=0.000 n=17+17)

name                      old alloc/op   new alloc/op   delta
InstrumentCreation-8        11.4kB ± 0%    10.1kB ± 0%  -10.97%  (p=0.000 n=20+19)

name                      old allocs/op  new allocs/op  delta
InstrumentCreation-8           270 ± 0%       222 ± 0%  -17.78%  (p=0.000 n=20+20)
```

### `instProviderKey` type

The `resolver` and `inserter` both accept a `view.Instrument` and `unit.Unit`. Unify these types into a single new `instProviderKey` that represents the new instrument request received by an instrument provider.

This enables easy caching for a future PR (see https://github.com/open-telemetry/opentelemetry-go/issues/3245)

### Store the `instrumentation.Scope` in the `insterter` type instead of the `meter`

The `meter` type currently has a field that holds the associated `instrumentation.Scope` of the meter.

https://github.com/open-telemetry/opentelemetry-go/blob/b5b685249cb2d2b8a3da72ed4d9c057ee6d0ba38/sdk/metric/meter.go#L33-L35

This field is only maintained to pass to an instrument provider as one of its fields, which in turn passes it to a `resolver` for each instrument creation, which then passes it to an `inserter`. Instead of maintaining state in all these types, just have the `inserter` define it as a field directly.