@MrAlias 2022-10-20
This doesn't close #3256, but it does block progress in adding caching to the instrument providers.